move logic to MintHTTPProvider
